:recycle: Move model dict to utilities.
