+++ importing/views.py
@@ -1,12 +1,11 @@
-from contextlib import suppress
-
 from django import forms
 from django.contrib.auth.decorators import login_required
-from django.db.models import ForeignKey, Model
 from django.http import HttpRequest, HttpResponse
 from django.shortcuts import get_object_or_404, render
 from django_tables2 import RequestConfig
 from guardian.shortcuts import get_objects_for_user
+
+from utilities.view_tools import model_to_dict
 
 from .models import DataImport
 from .tables import DataImportTable
@@ -33,29 +32,6 @@ def data_import_list(request: HttpRequest) -> HttpResponse:
         "table.html",
         {"table": table, "title": "Data imports", "refresh": "import:data_import_list"},
     )
-
-
-def model_to_dict(instance: Model) -> dict:
-    """Convert a model instance to a dictionary.
-
-    For ForeignKey fields, the related model instance is converted to a string, so that
-    a meaningful representation is shown in the dictionary instead of the primary key.
-
-    Args:
-        instance (Model): Model instance to convert.
-
-    Returns:
-        dict: Dictionary with the model instance data.
-    """
-    data = {}
-    for field in instance._meta.get_fields():
-        data[field.name] = field.value_from_object(instance)
-        if isinstance(field, ForeignKey):
-            with suppress(field.related_model.DoesNotExist):
-                data[field.name] = str(
-                    field.related_model.objects.get(pk=data[field.name])
-                )
-    return data
 
 
 @login_required
